+++ semana10/labex/src/App.css
@@ -5,7 +5,6 @@ flex-direction: column;
 height: 100vh;
 width: 100vw;
 align-content: center;
-background-color: black;
 /* margin: auto; */
 color:white;
 }
@@ -13,6 +12,7 @@ color:white;
 .App-logo {
   height: 40vmin;
   pointer-events: none;
+  
 }
 
 @media (prefers-reduced-motion: no-preference) {
