+++ semana10/labex/src/pages/TripDetailPage/TripDetailPage.js
@@ -1,24 +1,19 @@
 import React, { useState, useEffect } from "react";
 import axios from "axios";
 import { BASE_URL } from "../../constants/url";
-import { useHistory } from "react-router-dom";
 import useProtectedPage from "../../hooks/useProtectedPage";
-import DeleteIcon from "@material-ui/icons/Delete";
-import IconButton from "@material-ui/core/IconButton";
 import Button from "@material-ui/core/Button";
 import DescriptionIcon from "@material-ui/icons/Description";
 import AccountBoxIcon from "@material-ui/icons/AccountBox";
-// import { makeStyles } from '@material-ui/core/styles';
 import Footer from "../../components/Footer/Footer";
 import Header from "../../components/Header/Header";
-import { TripDetailContainer } from "../styledPages";
+import { TripDetailContainer } from "../../styledPages/styledPages";
 import ToApprovePage from "./ToApprovePage";
 import ApprovedPage from "./ApprovedPage";
 import { useParams } from "react-router-dom";
 
 const TripDetailPage = () => {
   const [tripDetail, setTripDetail] = useState();
-  const history = useHistory();
   const [approved, setApproved] = useState(false);
   const [toApprove, setToApprove] = useState(false);
   const params = useParams();
@@ -47,9 +42,9 @@ const TripDetailPage = () => {
       .catch((err) => alert(err.response.data));
   });
 
-  return (
-    <TripDetailContainer>
-      <Header />
+  return ( 
+    <TripDetailContainer>   
+            <Header />
       <div className="TripDetailBodyContainer">
         <h2> Detalhes da Viagem </h2>
         {tripDetail ? (
